Start moving classes to starter library
