--- sso-kit-hilla-demo/src/main/java/com/example/application/data/Role.java
@@ -1,5 +1,0 @@
-package com.example.application.data;
-
-public enum Role {
-    USER, ADMIN;
-}
