--- sso-kit-hilla-demo/src/main/java/com/example/application/data/service/UserRepository.java
@@ -1,10 +1,0 @@
-package com.example.application.data.service;
-
-import com.example.application.data.entity.User;
-import java.util.UUID;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface UserRepository extends JpaRepository<User, UUID> {
-
-    User findByUsername(String username);
-}
