--- sso-kit-hilla-demo/src/main/java/com/example/application/data/service/UserService.java
@@ -1,41 +1,0 @@
-package com.example.application.data.service;
-
-import com.example.application.data.entity.User;
-import java.util.Optional;
-import java.util.UUID;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
-import org.springframework.stereotype.Service;
-
-@Service
-public class UserService {
-
-    private final UserRepository repository;
-
-    @Autowired
-    public UserService(UserRepository repository) {
-        this.repository = repository;
-    }
-
-    public Optional<User> get(UUID id) {
-        return repository.findById(id);
-    }
-
-    public User update(User entity) {
-        return repository.save(entity);
-    }
-
-    public void delete(UUID id) {
-        repository.deleteById(id);
-    }
-
-    public Page<User> list(Pageable pageable) {
-        return repository.findAll(pageable);
-    }
-
-    public int count() {
-        return (int) repository.count();
-    }
-
-}
