--- sso-kit-hilla-starter/src/main/java/com/vaadin/sso/hilla/starter/Configuration.java
@@ -1,4 +1,0 @@
-package com.vaadin.sso.hilla.starter;
-
-public class Configuration {
-}
